Added git-edit script
When commiting and realizing that there's a file included in the commit
that belongs somewhere else, it's cumbersome to keep the log message.
git-edit does that: records the latest commit message, resets --soft,
then allows any changes. Finally, when the user shell is closed,
recommits the file with the previous commit message.