Added "imgshrink" script
Script I used to resize the ext4 partition on oversized disk images.
For example, they only need 1.3 GiB root fs, but the image file is
allocated to be 8 GiB -- it would easily fit a 4 GiB stick, but you
need to resize it first. This script determines the partition offset,
resizes the ext4 partition inside, changes the partition table and shows
you how much of the original image you need to dd on your disk.