Improve imgshrink a bit
Add more sanity checks to imgshrink, add some slack space as well as a
minimum amount of free space. Additionally, after modifying the
partition table, run resize2fs again to allow for the full extent of the
partition. Also be more conservative about the loop device mapping and
limit the mapped size to that of the actual partition so we do not
accidently overwrite following partition data.