Record only screen-on time for determination of task dates
When we want to run a backup periodically, it doesn't make sense to
categorically run it every 24h. We want to run it after maybe 6 hours of
actual *work* that has been done (i.e., "screensaver off time").
Therefore we now keep track of this metic and determine task runs
depending on that.