I hate Snap
Snaps are so insanely annoying. Not only update they with idiotically
high frequency, breaking software in the process, but also for minor
updates, not just security. Worse yet is the ultra-annoying notification
that cannot be shut off, apparently, under MATE. I need to update this
manuylly in the future because I am losing my mind otherwise.